backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.131.0",
    "google-genai>=1.64.0",
    "pandas>=3.0.1",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
    "markitdown>=0.1.2",
    "openpyxl>=3.1.5",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.41.0",
]
